requirements.txt
bqms-run @ file:///Users/ryandenotter/required_repos/dwh-migration-tools/client
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpathlib==0.10.0
config==0.5.1
exceptiongroup==1.1.1
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-bigquery==3.11.0
google-cloud-bigquery-migration==0.11.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
iniconfig==2.0.0
marshmallow==3.19.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
